Add formatDate helper
