+++ components/dashboard/GrantCard.tsx
@@ -1,23 +1,10 @@
 import Link from "next/link"
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
+import { formatDate } from "@/utils/helperFunctions";
 
 export default function GrantCard({grantID, name, amount, description, createdDate}: {grantID: number, name: string, amount: number, description: string, createdDate: string}){
-    const months = {
-        "01": "January",
-        "02": "February",
-        "03": "March",
-        "04": "April",
-        "05": "May",
-        "06": "June",
-        "07": "July",
-        "08": "August",
-        "09": "September",
-        "10": "October",
-        "11": "November",
-        "12": "December",
-    }
 
-    const date = months[createdDate.slice(5, 7)] + " " + createdDate.slice(8, 10) + ", " + createdDate.slice(0, 4);
+    const date = formatDate(createdDate);
     return(
         <div className="flex flex-col justify-between border-2 border-neutral-400 px-4 py-4 rounded-lg bg-[#FFFEFE] 2xl:w-[28rem] w-[20rem] 2xl:mr-6 mr-6 2xl:my-4 my-3 group hover:text-darkmaroon hover:border-darkmaroon transition duration-500">
             <div className="">
